refactor: restructure application to use ES modules and update routing
